Refactored deletion service methods into collection service
