+++ src/Aggregator/Core/Services/Contracts/IAggregationService.cs
@@ -16,7 +16,5 @@ namespace Biobanks.Aggregator.Core.Services.Contracts
         MaterialDetail GenerateMaterialDetail(IEnumerable<LiveSample> samples);
 
         string GenerateCollectionName(LiveSample sample);
-
-        Task DeleteMaterialDetailsBySampleSetId(int id);
     }
 }
