+++ src/Aggregator/Core/Services/Contracts/ICollectionService.cs
@@ -7,10 +7,14 @@ namespace Biobanks.Aggregator.Core.Services.Contracts
     {
         Task<Collection> GetCollectionAsync(int organisationId, string collectionName);
 
-        Task DeleteCollectionAsync(Collection collection);
-        
         Task UpdateCollectionAsync(Collection collection);
         
         Task AddCollectionAsync(Collection collection);
+
+        Task DeleteCollectionAsync(Collection collection);
+
+        Task DeleteSampleSetById(int id);
+
+        Task DeleteMaterialDetailsBySampleSetId(int id);
     }
 }
