+++ src/Aggregator/Core/Services/Contracts/ISampleService.cs
@@ -1,5 +1,4 @@
 ﻿using Biobanks.Entities.Api;
-using Biobanks.Entities.Data;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -15,6 +14,5 @@ namespace Biobanks.Aggregator.Core.Services.Contracts
 
         Task DeleteFlaggedSamplesAsync();
 
-        Task DeleteSampleSetById(int id);
     }
 }
